move LoginRequiredWrapper
