+++ src/pages/login/index.tsx
@@ -6,7 +6,7 @@ import styled from 'styled-components';
 
 import { signIn } from 'next-auth/client';
 
-import { LoginRequiredWrapper } from '~/components/Authentication/LogoutRequiredWrapper';
+import { LoginRequiredWrapper } from '~/components/common/Authentication/LogoutRequiredWrapper';
 import { WebevOgpHead } from '~/components/common/WebevOgpHead';
 import { imagePath } from '~/const/imagePath';
 import { toastSuccess } from '~/utils/toastr';
